Added separate file CommentBoxes.tsx
