+++ packages/web/app/userFeedback/page.tsx
@@ -1,7 +1,7 @@
 
 import { supabase } from '../../lib/supabase/supabaseClient';
 import ThreeCardLayout from '../../components/ThreeCardLayout';
-import NewComment from '@/components/Comment';
+
 
 export const dynamic = "force-dynamic";
